First Commit
A Minimum Viable Product

The game has been successfully tested manually, and several bugs were
discovered. These bugs are not critical; they have a minor impact on
the gameplay comfort but do not hinder the game's functionality.
Additionally, some code elements do not adhere to software design
principles such as SOLID (the game uses a single large class called
"Game" responsible for storing all game data and methods affecting this
data) and DRY (a small portion of the code, about 20-25 lines,
is duplicated). Moreover, there is a lot of commented-out, vestigial code in the
program that needs to be removed. Furthermore, the code lacks
sufficient comments explaining the functions' purposes and their
principles of operation.

Bugs to be fixed:
- Unnecessary screen refresh after another player purchases a card.
- Incorrect display of the panel showing the number of tokens owned by
the client_player during their turn.

Requires further improvement:
- Fixing the above-mentioned bugs.
- Adding a restriction on moves made out of turn.
- Writing unit tests.
- Removing vestigial code.
- Adding comments to functions, explaining their purposes and
principles of operation.